Adds authorization + login

+++ controllers/contactsControllers.js
@@ -3,7 +3,7 @@ import HttpError from "../helpers/HttpError.js";
 
 export const getAllContacts = async (req, res, next) => {
   try {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ ownerId: req.user.id });
 
     if (!contacts) throw HttpError(404);
 
@@ -20,6 +20,8 @@ export const getOneContact = async (req, res, next) => {
 
     if (!contact) throw HttpError(404);
 
+    if (contact.ownerId.toString() !== req.user.id) throw HttpError(403);
+
     res.status(200).json(contact);
   } catch (error) {
     next(error);
@@ -29,9 +31,12 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const contactId = req.params.id;
-    const contact = await Contact.findByIdAndDelete(contactId);
+    const contact = await Contact.findById(contactId);
 
     if (!contact) throw HttpError(404);
+    if (contact.ownerId.toString() !== req.user.id) throw HttpError(403);
+
+    await Contact.findByIdAndDelete(contactId);
 
     res.status(200).json(contact);
   } catch (error) {
@@ -42,7 +47,9 @@ export const deleteContact = async (req, res, next) => {
 export const createContact = async (req, res, next) => {
   try {
     const { name, email, phone } = req.body;
-    await Contact.create({ name, email, phone });
+    const ownerId = req.user.id;
+
+    await Contact.create({ name, email, phone, ownerId });
 
     res.status(201).json(Contact);
   } catch (error) {
@@ -53,14 +60,16 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
   try {
     const contactId = req.params.id;
+    const contact = await Contact.findById(contactId);
+
+    if (!contact) throw HttpError(404);
+    if (contact.ownerId.toString() !== req.user.id) throw HttpError(403);
 
     const updatedContact = await Contact.findByIdAndUpdate(
       contactId,
       req.body,
       { new: true }
     );
-
-    if (!updatedContact) throw HttpError(404);
 
     res.status(200).json(updatedContact);
   } catch (error) {
@@ -71,14 +80,16 @@ export const updateContact = async (req, res, next) => {
 export const updateStatusContact = async (req, res, next) => {
   try {
     const contactId = req.params.id;
+    const contact = await Contact.findById(contactId);
+
+    if (!contact) throw HttpError(404);
+    if (contact.ownerId.toString() !== req.user.id) throw HttpError(403);
 
     const updatedStatusContact = await Contact.findByIdAndUpdate(
       contactId,
       req.body,
       { new: true }
     );
-
-    if (!updatedStatusContact) throw HttpError(404);
 
     res.status(200).json(updatedStatusContact);
   } catch (error) {
